classes e calculadora

+++ primeiro programa java/src/com/dio/calculadora/Calculadora.java
@@ -17,7 +17,7 @@ public class Calculadora {
 		int soma = soma(a, b);
 		int subtracao = subtracao(a, b);
 		int multiplicacao = multiplicacao(a, b);
-		int divisao = divisao(a, b);
+		double divisao = divisao(a, b);
 
 		System.out.println(soma);
 		System.out.println(subtracao);
@@ -34,7 +34,7 @@ public class Calculadora {
 		return a - b;
 	}
 
-	public static int divisao(int a, int b) {
+	public static double divisao(double a, double b) {
 		return a / b;
 	}
